week01-llm-api/minseon/requirements.txt
openai
python-dotenv
streamlit
fastapi
uvicorn
